// src/apis/home/homeApi.model.ts
export type TotalContent = {
  presentation: number;
  presentationStore: number;
  asset: number;
  playlist: number;
};

export type TotalDevice = {
  total: number;
  online: number;
  play: number;
};

export type StorageData = {
  totalStorage: number;
  usedStorage: number;
  eCoinX: number;
  eCoinF: number;
  expiredDate: Date;
};